role

+++ src/controllers/product.controller.js
@@ -1,4 +1,6 @@
+const { ROLES, MESSAGES } = require('../constants')
 const productModel = require('../models/product.model')
+const userModel = require('../models/user.model')
 
 const getAll = async (req, res) => {
   try {
@@ -26,6 +28,12 @@ const getById = async (req, res) => {
 
 const create = async (req, res) => {
   try {
+    const profile = await userModel.findUserById(req.user.id)
+    const allowedRoles = [ROLES.ADMIN, ROLES.SELLER]
+
+    if (!allowedRoles.includes(profile.role_name))
+      return res.status(403).json({ message: MESSAGES.UNAUTHORIZED })
+
     const imageUrls = req.files?.map(file => `/uploads/${file.filename}`) || []
     const product = await productModel.createProduct(req.body, imageUrls)
     res.status(201).json(product)
@@ -38,7 +46,15 @@ const create = async (req, res) => {
 
 const update = async (req, res) => {
   try {
+    const profile = await userModel.findUserById(req.user.id)
+    const allowedRoles = [ROLES.ADMIN, ROLES.SELLER]
+
+    if (!allowedRoles.includes(profile.role_name))
+      return res.status(403).json({ message: MESSAGES.UNAUTHORIZED })
+
     const imageUrls = req.files?.map(file => `/uploads/${file.filename}`) || []
+    console.log('imageUrls', imageUrls)
+
     const product = await productModel.updateProduct(
       req.params.id,
       req.body,
@@ -53,8 +69,18 @@ const update = async (req, res) => {
 }
 
 const remove = async (req, res) => {
-  await productModel.deleteProduct(req.params.id)
-  res.json({ message: 'Đã xoá sản phẩm' })
+  try {
+    const profile = await userModel.findUserById(req.user.id)
+    const allowedRoles = [ROLES.ADMIN, ROLES.SELLER]
+
+    if (!allowedRoles.includes(profile.role_name))
+      return res.status(403).json({ message: MESSAGES.UNAUTHORIZED })
+
+    await productModel.deleteProduct(req.params.id)
+    res.json({ message: 'Đã xoá sản phẩm' })
+  } catch (err) {
+    res.status(500).json({ message: 'Lỗi xóa sản phẩm', error: err.message })
+  }
 }
 
 module.exports = { getAll, getById, create, update, remove }
